requirements.txt
django-material
django-smartfields